+++ src/Qlarissa.Infrastructure/DB/Entities/MarketData/DividendPayout.cs
@@ -10,7 +10,7 @@ public sealed class DividendPayout
 
     public DateOnly PayoutDate { get; set; }
 
-    public decimal PayoutAmount { get; set; }
+    public double PayoutAmount { get; set; }
 
     public int SecurityId { get; set; }
